Reworked
Removed a lot of fluff.
Got rid of querys that really had no purpose, now it takes the file and translates it to amino acids whether it's RNA or DNA. I created separate variables to handle pre and post translation.

Removed batch iteration

Added Carbon Count, currently working on logistics of that.